Integrate new system of folders distribution
