+++ src/legacy/photosProcessingCycle.cpp
