+++ src/legacy/photosProcessingCycle.h
