+++ src/legacy/videoProcessingCycle.cpp
